Move all routing and navigation related classed to their corresponding namespace
