+++ src/Sulu/Bundle/AdminBundle/Admin/Admin.php
@@ -11,6 +11,11 @@
 
 namespace Sulu\Bundle\AdminBundle\Admin;
 
+use Sulu\Bundle\AdminBundle\Admin\Navigation\NavigationItemCollection;
+use Sulu\Bundle\AdminBundle\Admin\Navigation\NavigationProviderInterface;
+use Sulu\Bundle\AdminBundle\Admin\Routing\RouteCollection;
+use Sulu\Bundle\AdminBundle\Admin\Routing\RouteProviderInterface;
+
 /**
  * Defines all the required information from a bundle's admin class.
  */
