+++ src/Sulu/Bundle/AdminBundle/Admin/Navigation/NavigationProviderInterface.php
@@ -9,7 +9,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\AdminBundle\Admin;
+namespace Sulu\Bundle\AdminBundle\Admin\Navigation;
 
 interface NavigationProviderInterface
 {
